requirements.txt
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
future==0.18.2
idna==3.4
oauthlib==3.2.2
Pillow==9.3.0
pycodestyle==2.10.0
python-dotenv==0.21.0
python-twitter==3.5
requests==2.28.1
requests-oauthlib==1.3.1
tomli==2.0.1
urllib3==1.26.13
